Rewrite memory allocation classes
